chore: move profile setting files
